Moving most ugly impl details from scala.pickling into scala.pickling.internal
