+++ core/src/test/scala/pickling/file-io.scala
@@ -5,6 +5,7 @@ import java.io.File
 
 import scala.io.Source
 import scala.pickling._
+import scala.pickling.internal._
 import scala.pickling.json._
 import scala.pickling.io.TextFileOutput
